Run the fbgemm block FP8 fallback tests off sm_90 for PR #8953
The module level skip keyed on compute capability major == 9, so all four
tests skipped everywhere else. Three of them take the dequant + matmul
fallback and never call fbgemm, so only the kernel battery needs the gate,
and gating it on unsloth's own import time probe rather than sm_90 lets
future arches enable themselves. Verified on B200 (sm_100): 3 passed,
1 skipped, versus 4 skipped before.

Also cast the f8f8bf16 output to X.dtype so the kernel and fallback
branches agree (no-op for bf16), and detach in _rel_err.

+++ unsloth/kernels/fp8.py
@@ -582,9 +582,10 @@ class FP8_fbgemm_block_linear(torch.autograd.Function):
         with _fp8_triton_device_context(X):
             # X is (tokens, K): its blocks are (bs_m, bs_k), not (bs_m, bs_n).
             xq, xs = triton_quantize_fp8_block(X, bs_m, bs_k, None)
+        # f8f8bf16 always returns bf16; cast so both branches return X.dtype.
         output = torch.ops.fbgemm.f8f8bf16_blockwise(
             xq, weight.contiguous(), xs, weight_scale.contiguous(), bs_m, bs_n, bs_k
-        )
+        ).to(X.dtype)
         output = output + bias if bias is not None else output
 
         output = output.view(*orig_shape[:-1], -1)
